Fixed bad cursor behaviour in IE
This fixes #3 where IE was moving the cursor to the end of the document
when editing text in the middle of the document. The fix was to check in
the EditorContent class in the text method to check if the text that is
being set is the same as what is already in the textarea; this should
also cut down on the chatiness of the events.

Also updated the demo files to point to the latest version of the files.