Setup bower and NPM packages
Now the library has the ability to be distributed via either bower or
npm. This also means that the distribution files get committed to the
repository.

Due to bower having an existing commonmark-editor package the library
has been renamed to bootstrap-commonmark-editor which is more accurate
to what the library is.

Closes #9
Closes #10